Fix bugs from merging develop
